+++ test/distance/tests-LCSseq.cpp
@@ -1,20 +1,62 @@
+#include "rapidfuzz/distance/LCSseq.hpp"
 #include <catch2/catch_approx.hpp>
 #include <catch2/catch_test_macros.hpp>
 #include <string>
 
 #include <rapidfuzz/distance/Indel.hpp>
 
-TEST_CASE("Indel")
+template <typename Sentence1, typename Sentence2>
+int64_t lcs_seq_distance(const Sentence1& s1, const Sentence2& s2,
+                         int64_t max = std::numeric_limits<int64_t>::max())
+{
+    int64_t res1 = rapidfuzz::lcs_seq_distance(s1, s2, max);
+    rapidfuzz::CachedLCSseq<typename Sentence1::value_type> scorer(s1);
+    int64_t res2 = scorer.distance(s2, max);
+    REQUIRE(res1 == res2);
+    return res1;
+}
+
+template <typename Sentence1, typename Sentence2>
+int64_t lcs_seq_similarity(const Sentence1& s1, const Sentence2& s2, int64_t max = 0)
+{
+    int64_t res1 = rapidfuzz::lcs_seq_similarity(s1, s2, max);
+    rapidfuzz::CachedLCSseq<typename Sentence1::value_type> scorer(s1);
+    int64_t res2 = scorer.similarity(s2, max);
+    REQUIRE(res1 == res2);
+    return res1;
+}
+
+template <typename Sentence1, typename Sentence2>
+double lcs_seq_normalized_distance(const Sentence1& s1, const Sentence2& s2, double score_cutoff = 1.0)
+{
+    double res1 = rapidfuzz::lcs_seq_normalized_distance(s1, s2, score_cutoff);
+    rapidfuzz::CachedLCSseq<typename Sentence1::value_type> scorer(s1);
+    double res2 = scorer.normalized_distance(s2, score_cutoff);
+    REQUIRE(res1 == Catch::Approx(res2).epsilon(0.0001));
+    return res1;
+}
+
+template <typename Sentence1, typename Sentence2>
+double lcs_seq_normalized_similarity(const Sentence1& s1, const Sentence2& s2, double score_cutoff = 0.0)
+{
+    double res1 = rapidfuzz::lcs_seq_normalized_similarity(s1, s2, score_cutoff);
+    rapidfuzz::CachedLCSseq<typename Sentence1::value_type> scorer(s1);
+    double res2 = scorer.normalized_similarity(s2, score_cutoff);
+    REQUIRE(res1 == Catch::Approx(res2).epsilon(0.0001));
+    return res1;
+}
+
+TEST_CASE("LCSseq")
 {
     std::string test = "aaaa";
     std::string replace_all = "bbbb";
 
     SECTION("similar strings")
     {
-        REQUIRE(rapidfuzz::lcs_seq_distance(test, test) == 0);
-        REQUIRE(rapidfuzz::lcs_seq_similarity(test, test) == 4);
-        REQUIRE(rapidfuzz::lcs_seq_normalized_distance(test, test) == 0.0);
-        REQUIRE(rapidfuzz::lcs_seq_normalized_similarity(test, test) == 1.0);
+        REQUIRE(lcs_seq_distance(test, test) == 0);
+        REQUIRE(lcs_seq_similarity(test, test) == 4);
+        REQUIRE(lcs_seq_normalized_distance(test, test) == 0.0);
+        REQUIRE(lcs_seq_normalized_similarity(test, test) == 1.0);
     }
 
     SECTION("completly different strings")
@@ -29,21 +71,21 @@ TEST_CASE("Indel")
     {
         std::string a = "South Korea";
         std::string b = "North Korea";
-        REQUIRE(rapidfuzz::lcs_seq_distance(a, b) == 2);
-        REQUIRE(rapidfuzz::lcs_seq_distance(a, b, 4) == 2);
-        REQUIRE(rapidfuzz::lcs_seq_distance(a, b, 3) == 2);
-        REQUIRE(rapidfuzz::lcs_seq_distance(a, b, 2) == 2);
-        REQUIRE(rapidfuzz::lcs_seq_distance(a, b, 1) == 2);
-        REQUIRE(rapidfuzz::lcs_seq_distance(a, b, 0) == 1);
+        REQUIRE(lcs_seq_distance(a, b) == 2);
+        REQUIRE(lcs_seq_distance(a, b, 4) == 2);
+        REQUIRE(lcs_seq_distance(a, b, 3) == 2);
+        REQUIRE(lcs_seq_distance(a, b, 2) == 2);
+        REQUIRE(lcs_seq_distance(a, b, 1) == 2);
+        REQUIRE(lcs_seq_distance(a, b, 0) == 1);
 
         a = "aabc";
         b = "cccd";
-        REQUIRE(rapidfuzz::lcs_seq_distance(a, b) == 3);
-        REQUIRE(rapidfuzz::lcs_seq_distance(a, b, 4) == 3);
-        REQUIRE(rapidfuzz::lcs_seq_distance(a, b, 3) == 3);
-        REQUIRE(rapidfuzz::lcs_seq_distance(a, b, 2) == 3);
-        REQUIRE(rapidfuzz::lcs_seq_distance(a, b, 1) == 2);
-        REQUIRE(rapidfuzz::lcs_seq_distance(a, b, 0) == 1);
+        REQUIRE(lcs_seq_distance(a, b) == 3);
+        REQUIRE(lcs_seq_distance(a, b, 4) == 3);
+        REQUIRE(lcs_seq_distance(a, b, 3) == 3);
+        REQUIRE(lcs_seq_distance(a, b, 2) == 3);
+        REQUIRE(lcs_seq_distance(a, b, 1) == 2);
+        REQUIRE(lcs_seq_distance(a, b, 0) == 1);
     }
 
     SECTION("testCachedImplementation")