+++ test/tests-fuzz.cpp
@@ -6,34 +6,31 @@
 namespace fuzz = rapidfuzz::fuzz;
 using Catch::Approx;
 
-using MetricPtr = double(*)(const char*, const char*, double);
-struct Metric
-{
+using MetricPtr = double (*)(const char*, const char*, double);
+struct Metric {
     MetricPtr call;
     const char* name;
     bool symmetric;
 };
 
-#define LIST_OF_METRICS(FUNC) \
-    /*   func                          symmetric */ \
-    FUNC(fuzz::ratio,                    true) \
-    FUNC(fuzz::partial_ratio,            false) \
-    FUNC(fuzz::token_set_ratio,          true)  \
-    FUNC(fuzz::token_sort_ratio,         true)  \
-    FUNC(fuzz::token_ratio,              true)  \
-    FUNC(fuzz::partial_token_set_ratio,  false) \
-    FUNC(fuzz::partial_token_sort_ratio, false) \
-    FUNC(fuzz::partial_token_ratio,      false) \
-    FUNC(fuzz::WRatio,                   false) \
-    FUNC(fuzz::QRatio,                   true)
+#define LIST_OF_METRICS(FUNC)                                                                                \
+    /*   func                          symmetric */                                                          \
+    FUNC(fuzz::ratio, true)                                                                                  \
+    FUNC(fuzz::partial_ratio, false)                                                                         \
+    FUNC(fuzz::token_set_ratio, true)                                                                        \
+    FUNC(fuzz::token_sort_ratio, true)                                                                       \
+    FUNC(fuzz::token_ratio, true)                                                                            \
+    FUNC(fuzz::partial_token_set_ratio, false)                                                               \
+    FUNC(fuzz::partial_token_sort_ratio, false)                                                              \
+    FUNC(fuzz::partial_token_ratio, false)                                                                   \
+    FUNC(fuzz::WRatio, false)                                                                                \
+    FUNC(fuzz::QRatio, true)
 
-#define CREATE_METRIC(func, symmetric) Metric{ \
-    [](const char* s1, const char* s2, double score_cutoff){ return func(s1, s2, score_cutoff); }, \
-    #func, symmetric},
+#define CREATE_METRIC(func, symmetric)                                                                       \
+    Metric{[](const char* s1, const char* s2, double score_cutoff) { return func(s1, s2, score_cutoff); },   \
+           #func, symmetric},
 
-std::vector<Metric> metrics = {
-    LIST_OF_METRICS(CREATE_METRIC)
-};
+std::vector<Metric> metrics = {LIST_OF_METRICS(CREATE_METRIC)};
 
 /**
  * @name RatioTest
@@ -116,29 +113,27 @@ TEST_CASE("RatioTest")
         REQUIRE(100 == fuzz::ratio("", ""));
         REQUIRE(100 == fuzz::partial_ratio("", ""));
         REQUIRE(100 == fuzz::token_sort_ratio("", ""));
-        REQUIRE(  0 == fuzz::token_set_ratio("", ""));
+        REQUIRE(0 == fuzz::token_set_ratio("", ""));
         REQUIRE(100 == fuzz::partial_token_sort_ratio("", ""));
-        REQUIRE(  0 == fuzz::partial_token_set_ratio("", ""));
+        REQUIRE(0 == fuzz::partial_token_set_ratio("", ""));
         REQUIRE(100 == fuzz::token_ratio("", ""));
         REQUIRE(100 == fuzz::partial_token_ratio("", ""));
-        REQUIRE(  0 == fuzz::WRatio("", ""));
-        REQUIRE(  0 == fuzz::QRatio("", ""));
+        REQUIRE(0 == fuzz::WRatio("", ""));
+        REQUIRE(0 == fuzz::QRatio("", ""));
     }
 
     SECTION("testFirstStringEmpty")
     {
-        for (auto& metric : metrics)
-        {
-            INFO( "Score not 0 for " << metric.name );
+        for (auto& metric : metrics) {
+            INFO("Score not 0 for " << metric.name);
             REQUIRE(0 == metric.call("test", "", 0));
         }
     }
 
     SECTION("testSecondStringEmpty")
     {
-        for (auto& metric : metrics)
-        {
-            INFO( "Score not 0 for " << metric.name );
+        for (auto& metric : metrics) {
+            INFO("Score not 0 for " << metric.name);
             REQUIRE(0 == metric.call("", "test", 0));
         }
     }
@@ -153,10 +148,9 @@ TEST_CASE("RatioTest")
         const char* str1 = "South Korea";
         const char* str2 = "North Korea";
 
-        for (auto& metric : metrics)
-        {
+        for (auto& metric : metrics) {
             double score = metric.call(str1, str2, 0);
-            INFO( "score_cutoff does not work correctly for " << metric.name );
+            INFO("score_cutoff does not work correctly for " << metric.name);
             REQUIRE(0 == metric.call(str1, str2, score + 0.0001));
             REQUIRE(score == metric.call(str1, str2, score));
         }
@@ -167,10 +161,9 @@ TEST_CASE("RatioTest")
         const char* str1 = "bc";
         const char* str2 = "bca";
 
-        for (auto& metric : metrics)
-        {
+        for (auto& metric : metrics) {
             double score = metric.call(str1, str2, 0);
-            INFO( "score_cutoff does not work correctly for " << metric.name );
+            INFO("score_cutoff does not work correctly for " << metric.name);
             REQUIRE(0 == metric.call(str1, str2, score + 0.0001));
             REQUIRE(score == metric.call(str1, str2, score));
         }
@@ -178,8 +171,10 @@ TEST_CASE("RatioTest")
 
     SECTION("testIssue231") /* test for https://github.com/maxbachmann/RapidFuzz/issues/231 */
     {
-        const char str1[] = "er merkantilismus f/rderte handel und verkehr mit teils marktkonformen, teils dirigistischen ma_nahmen.";
-        const char str2[] = "ils marktkonformen, teils dirigistischen ma_nahmen. an der schwelle zum 19. jahrhundert entstand ein neu";
+        const char str1[] = "er merkantilismus f/rderte handel und verkehr mit teils marktkonformen, teils "
+                            "dirigistischen ma_nahmen.";
+        const char str2[] = "ils marktkonformen, teils dirigistischen ma_nahmen. an der schwelle zum 19. "
+                            "jahrhundert entstand ein neu";
 
         auto alignment = fuzz::partial_ratio_alignment(str1, str2);
         REQUIRE(alignment.src_start == 0);
@@ -188,6 +183,3 @@ TEST_CASE("RatioTest")
         REQUIRE(alignment.dest_end == 103);
     }
 }
-
-
-