+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### [1.0.4] - 2022-
 #### Fixed
 - fix implementation of hamming_normalized_similarity
+- fix implementation of CachedLCSseq::distance
+
 ### [1.0.3] - 2022-06-24
 #### Fixed
 - fix integer wraparound in partial_ratio/partial_ratio_alignment